changed extract_amount to keep the value as a string
reading CSVs is a problem, because it just reads it as a dictionary and python automatically makes them all strings. It would take work that I don't want to do right now to parse each type from the string for each column in each row. So for now if somethign needs to be converted to a type, it will just have to be done in place.