refactored create_and_write_dict_list to write_dict_list
it now contains the logic for determining if the file already exists and simply adds new rows to it. if the file doesn't exist it creates the new file, the header, and then adds the rows to it. also we are no longer using the write_dict.keys() to determine the fieldname parameter. It was not matching things up, and only worked in a specific order. For this reason I created a new list transaction_write_list and it holds the column names that match up with incoming dicts to be written